Refactor: header 리팩토링
Design: header 디자인 수정
